Add bind function
Add a bind function for the new union type. The function should take two arguments:
The first argument to the bind function should be a function that takes the generic type argument as input, and returns a value of the 'program' union type as output. In the rest of this recipe, this function is called f.
The second argument to the bind function should be a 'program' union type value.
The return type of the bind function should be a 'program' union type value, with the same generic type as the return type of the first argument (f).
Declare the bind function as recursive by adding the rec keyword.
Implement the bind function by pattern-matching on the Free and Pure cases:
In the Free case, pipe the contained functor value to the functor's map function, using bind f as the mapper function; then pipe the result of that to Free.
In the Pure case, return f x, where x is the value contained in the Pure case.